// src/app/components/shared/models/course.ts
export interface IObject {
    id: number;
    name: string;
}

export interface ICourse {
    COURSE_ID: number;
    COURSE_NAME: string;
    COURSE_DESCRIPTION: string;
    COURSE_CATEGORY_ID: number;
    COURSE_STATUS: boolean;
    COURSE_RATING: number;
    COURSE_IMAGE: string;
    IMAGE_URL: string;
    USERID: number;
    FIRSTNAME: string;
    LASTNAME: string;
    COURSE_TRENDING: boolean;
    COURSESHD_AMOUNT: number;
}